Move Highlighter to Seacher
